+++ routes/profile.js
@@ -4,11 +4,14 @@ const router = express.Router();
 const userData = require('../models/user');
 
 router.get('/:username', (req, res) => {
-    const user = userData.getByUsername(req.params.username);
-    if(user)
-        res.render('user/profile', { title: `${user.username}'s profile`, user: user });
-    else
-        res.render('not_found', { title: 'Not found'});
+    userData.get().then(data => {
+        res.send(data);
+    });
+    // const user = userData.getByUsername(req.params.username);
+    // if(user)
+    //     res.render('user/profile', { title: `${user.username}'s profile`, user: user });
+    // else
+    //     res.render('not_found', { title: 'Not found'});
 });
 
 module.exports = router;